refactored writers to make them consistent
