+++ src/main/scala/beam/utils/csv/conversion/XmlConverter.scala
@@ -1,4 +1,4 @@
-package beam.utils.csv
+package beam.utils.csv.conversion
 
 import java.io.{File, FileInputStream, FileOutputStream}
 import java.nio.file.Files
