+++ src/main/scala/beam/utils/csv/writers/ScenarioCsvWriter.scala
@@ -1,4 +1,4 @@
-package beam.utils.csv
+package beam.utils.csv.writers
 
 import java.io.File
