+++ src/test/scala/beam/utils/NetworkXmlToCsvSpec.scala
@@ -1,6 +1,6 @@
 package beam.utils
 
-import beam.utils.csv.NetworkXmlToCSV
+import beam.utils.csv.conversion.NetworkXmlToCSV
 import org.scalatest.WordSpecLike
 
 class NetworkXmlToCsvSpec extends WordSpecLike {
